feat: add CPU and vCPU fields inside the BaoIoRequest
This patch is based on passing the CPU and vCPU ID in the request structure itself since each client (VirtIO frontend) can be running on multiple CPUs and vCPUs. This way, when we pass on this information w have exact information about which order is being processed or not and which order should be completed. Without these fields, we could, for example, have limit cases where we would be completing requests from other vCPUs.

Signed-off-by: joaopeixoto13 <joaopeixotooficial@gmail.com>